Consistantly renamed Documentation components
